fix(aic8800): close post-rebase driver audit gaps
