+++ apps/starry/iperf3/board-aka-00-sg2002.session-env.toml
@@ -1,5 +1,1 @@
-inject_boot_entropy = true
-
 [wifi]
-ssid_file = "credentials/wifi-ssid"
-pmk_file = "credentials/wifi-pmk"
