+++ scripts/axbuild/src/starry/app/board_assets.rs
@@ -4,13 +4,13 @@ use anyhow::Context;
 
 use super::StarryAppBoardCase;
 use crate::{
-    starry::test::{
-        PreparedBoardSessionAssets, SessionAssetDelivery, SessionRunDirectoryGuard,
-        collect_upload_paths, copy_declared_session_files, starry_case_asset_config,
+    starry::{
+        board_assets::{BoardSessionAssetPlan, PreparedBoardSessionAssets},
+        test::starry_case_asset_config,
     },
     test::{
         build::{prepare_c_case_overlay_sync, prepare_rust_case_overlay_sync},
-        case::{TestQemuCase, board_case_asset_layout, reset_dir},
+        case::reset_dir,
     },
 };
 
@@ -23,16 +23,16 @@ pub(in crate::starry) async fn prepare_app_board_session_assets(
 ) -> anyhow::Result<Option<PreparedBoardSessionAssets>> {
     let rust_manifest = case.case_dir.join("rust/Cargo.toml");
     let c_manifest = case.case_dir.join("c/CMakeLists.txt");
-    let session_env = crate::starry::session_env::SessionEnvConfig::load(
+    let session_environment = crate::starry::session_env::SessionEnvConfig::load(
         &case.case_dir,
         &case.board_config_path,
-    )?;
-    if let Some(session_env) = &session_env {
-        session_env.validate_environment()?;
-    }
+    )?
+    .map(|config| config.capture_environment())
+    .transpose()?
+    .unwrap_or_default();
     if !rust_manifest.is_file()
         && !c_manifest.is_file()
-        && session_env.is_none()
+        && !session_environment.is_bootstrap()
         && declared_session_files.is_empty()
     {
         return Ok(None);
@@ -54,52 +54,39 @@ pub(in crate::starry) async fn prepare_app_board_session_assets(
     let case_dir = case.case_dir.clone();
     let board_config_path = case.board_config_path.clone();
     let declared_session_files = declared_session_files.to_vec();
-    let delivery = SessionAssetDelivery::for_session_env(session_env.as_ref());
-    let session_env = session_env.unwrap_or_default();
+    let plan = BoardSessionAssetPlan {
+        workspace_root,
+        target,
+        work_name: format!("app/{case_name}"),
+        case_name,
+        case_dir,
+        board_config_path,
+        declared_session_files,
+        session_environment,
+    };
 
     let assets = tokio::task::spawn_blocking(move || -> anyhow::Result<_> {
-        let layout =
-            board_case_asset_layout(&workspace_root, &target, &format!("app/{case_name}"))?;
-        let cleanup = SessionRunDirectoryGuard::new(layout.run_dir.clone());
-        let build_case = TestQemuCase {
-            name: case_name.clone(),
-            display_name: case_name,
-            case_dir: case_dir.clone(),
-            qemu_config_path: board_config_path,
-            test_commands: Vec::new(),
-            host_symbolize_success_regex: Vec::new(),
-            host_http_server: None,
-            subcases: Vec::new(),
-            grouped_subcase_filter: None,
-        };
-        if rust_manifest.is_file() {
-            prepare_rust_case_overlay_sync(
-                &arch,
-                &build_case,
-                rootfs.as_ref().expect("Rust pipeline requested a rootfs"),
-                &layout,
-                &starry_case_asset_config(),
-            )?;
-        } else if c_manifest.is_file() {
-            prepare_c_case_overlay_sync(
-                &arch,
-                &build_case,
-                rootfs.as_ref().expect("C pipeline requested a rootfs"),
-                &layout,
-                &starry_case_asset_config(),
-            )?;
-        } else {
-            reset_dir(&layout.overlay_dir)?;
-        }
-        copy_declared_session_files(&case_dir, &layout.overlay_dir, &declared_session_files)?;
-        session_env.materialize(&layout.overlay_dir)?;
-        let relative_paths = collect_upload_paths(&layout.overlay_dir)?;
-        Ok(PreparedBoardSessionAssets::new(
-            layout.overlay_dir,
-            relative_paths,
-            cleanup.preserve(),
-            delivery,
-        ))
+        plan.prepare(|build_case, layout| {
+            if rust_manifest.is_file() {
+                prepare_rust_case_overlay_sync(
+                    &arch,
+                    build_case,
+                    rootfs.as_ref().expect("Rust pipeline requested a rootfs"),
+                    layout,
+                    &starry_case_asset_config(),
+                )
+            } else if c_manifest.is_file() {
+                prepare_c_case_overlay_sync(
+                    &arch,
+                    build_case,
+                    rootfs.as_ref().expect("C pipeline requested a rootfs"),
+                    layout,
+                    &starry_case_asset_config(),
+                )
+            } else {
+                reset_dir(&layout.overlay_dir)
+            }
+        })
     })
     .await
     .context("Starry app board session asset task failed")??;
